refactor: address multi-SAE PR review feedback
- Share InterruptedException/interrupt_callback in sae_lens/training/_interruption
- Construct SAEs directly from configs instead of from_dict round-trip
- Replace hasattr duck-typing with isinstance(PrefetchingIterator) for paused()
- Drop the None sentinel on _hook_head_indices; always a dict
- Remove unnecessary torch.zeros allocation in get_multi_hook_activations
- Move _FakeWandbArtifact / captured_wandb_logs fixture to tests/conftest.py
- Strip docstrings from test functions per CLAUDE.md
